pyproject.toml
[project]
name = "awada"
version = "0.1.0"
description = "AWADA: Foreground-focused adversarial learning for cross-domain object detection"
requires-python = ">=3.8"
dependencies = [
    "torch>=1.13.0",
    "torchvision>=0.14.0",
    "numpy>=1.21.0",
    "Pillow>=9.0.0",
    "scipy>=1.7.0",
    "tqdm>=4.62.0",
    "pycocotools>=2.0.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
strict = false
